recommendation-service/requirements.txt
# FastAPI 관련
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ML 라이브러리 (ALS 모델용) - 코랩 환경 버전
implicit==0.7.2              # ALS 모델 라이브러리
pandas==2.2.2                # 코랩 환경과 맞춤
numpy==2.0.2                 # 코랩 환경과 맞춤
scipy==1.15.3                # 코랩 환경과 맞춤
scikit-learn==1.6.1          # 코랩 환경과 맞춤

# 데이터베이스 연결 (MySQL)
PyMySQL==1.1.0
SQLAlchemy==2.0.23
cryptography==41.0.7         # PyMySQL 암호화 지원

# HTTP 클라이언트
httpx==0.25.0
requests==2.31.0

# 캐싱 (선택사항)
redis==5.0.1

# 설정 파일 처리
PyYAML==6.0.1

# 유틸리티
python-multipart==0.0.6
python-json-logger==2.0.7

# 추가 호환성을 위한 패키지
wheel>=0.37.0

# Schedule dependency for batch processing
schedule==1.2.0